Add native python type support

+++ src/main/python/dynamic_hosting/example_model_training/miniloan-rfc.py
@@ -101,7 +101,7 @@ def main():
             {
                 'name': "creditScore",
                 'order': 0,
-                'type': 'float64'
+                'type': 'int'
             },
             {
                 'name': "income",
